Add quantized-mouse shader
This does two things: it quantizes the mouse to display a square in 20px increments, like they're snapped to a grid. It also shifts the color of the square based on the x/y position.

It does the quantization by dividing and floor-ing the mouse position by the desired grid value and multiplying it back by the same value so it looses precision. Then it's shifted by half of the grid value so the grid square is centered on the mouse.

The color is accomplished by using the current pixel which is the current coordinate / the total resolution to get a 0 - 1 range. Various x and y values are then used to set the color of the square.