Add sunset shader
This uses multiple mixes and smoothstep to "layer" gradients one after another. First it does the blue into the red, then takes the output of that and layers on the orange, then takes that and layers on the yellow.

I got the idea from here: http://stackoverflow.com/questions/15935117/how-to-create-multiple-stop-gradient-fragment-shader#15936103