Redo quantized mouse without a conditional
This uses a similar approach but uses various `step` functions to
subtract areas from the drawn canvas.

First if turns the mouse position into 20px quantized, square fragments
and shifts it by 20px in each direction. This shift allows the actual
mouse to always be inside of the eventual square instead of just outside
it. This is the same thing that's done in the original quantized mouse.

Then, it uses a series of steps to calculates whether or not the current
pixel is on some side of the mouse or another. It starts with all pixels
to the left of the quantized mouse. Then finds all pixels to the left of
that same point but one 20px segment to the left and subtracts that,
creating a 20px wide vertical line.

Then it finds all pixels below the current quantized mouse point
(shifted up 20px) and subtracts it if it is. Finally it subtracts all
current pixels above the quantized point and subtracts them.

After all that's done, it's left with a square of pixels that haven't
been subtracted. The original mouse coordinate is used to color the
square.